// Front/ProcesoMatricula/src/app/model/cursos.ts
export interface Cursos {
    id: number;
    planEstudioId: number;
    nombre: string;
    codigo: string;
    docenteId: number;
    horario: string;
    aula: string;
    periodo: string;
    ofertaAcademicaid?: number;
    FechaDeCreacion?: string;
    ActualizadoEn?: string;
  }

  /*
export interface Cursos {
  id: number;
  planEstudioId: number;
  nombre: string;
  codigo: string;
  docenteId: number;
  horario: string;
  aula: string;
  periodo: string;
  ofertaAcademicaid?: number;
  FechaDeCreacion?: string;
  ActualizadoEn?: string;
}
  */